Flounder wrapper rejiggery #60
